+++ core/src/db/DBImpl.cpp
@@ -694,7 +694,11 @@ DBImpl::ListIDInSegment(const std::string& collection_name, int64_t segment_id, 
     STATUS_CHECK(segment_reader->LoadDeletedDocs(deleted_docs_ptr));
     if (deleted_docs_ptr) {
         const std::vector<offset_t>& delete_ids = deleted_docs_ptr->GetDeletedDocs();
-        for (auto offset : delete_ids) {
+        std::vector<offset_t> temp_ids;
+        temp_ids.reserve(delete_ids.size());
+        std::copy(delete_ids.begin(), delete_ids.end(), std::back_inserter(temp_ids));
+        std::sort(temp_ids.begin(), temp_ids.end(), std::greater<offset_t>());
+        for (auto offset : temp_ids) {
             entity_ids.erase(entity_ids.begin() + offset, entity_ids.begin() + offset + 1);
         }
     }